refactor(#5031): extract TmuxWatcherRegistry impl (W5a)
